Triplets comparison - passed
This was an easy one but it made me remember a few of clojure core
functions. the code is not super generic but it was just enough to pass
the unit tests I've done before starting the actual code. I found that
TDD works best for educational purposes: when you're trying to implement
a well defined algorithm for example, something from the books.