Min-max sum completed
Had to recall the backtrack algorithm but in the end it's not even that.
It's just a simple recursive call (cause it's easier than loops in clojure).
This time I've tried to do some refactoring and created small helper functions.
They are probably not the perfect funcitons in the world and I'm sure the solution
can be written in a more streamlined way. But at this moment this should do.

Had to rewrite some functions in order to not only work for vectors but for sequences.
So I've learned about a couple of new functions.

It seems that I don't need a helper function to print the result. I can just apply the print
function to the collection. Cool.