Birthday Cake Candles complete
Decided to try and use reduce function instead of a loop this time.
Since we are reducing a collection of things to one thing - reduce seems the perfect fit.
But I needed to store the current max value somehow. So I've used a hash as a reduction point.
Worked nicely. I could first code in the data type that I wanted to work with {:max 0 :count 0}
and then just wrote the function that decided on the new values for both keys depend on the current
state and the new number.